+++ src/main/java/com/elmakers/mine/bukkit/traders/WandBoundFlag.java
@@ -2,7 +2,7 @@ package com.elmakers.mine.bukkit.traders;
 
 import net.dandielo.citizens.traders_v3.utils.items.Attribute;
 
-@Attribute(name="Wand Bound Status", key="wand_bound", priority = 5)
+@Attribute(name="Wand Bound Status", key=".wand_bound", priority = 5)
 public class WandBoundFlag extends NBTItemFlag {
 	
 	public WandBoundFlag(String key) {