+++ src/main/java/com/elmakers/mine/bukkit/traders/WandFillFlag.java
@@ -2,7 +2,7 @@ package com.elmakers.mine.bukkit.traders;
 
 import net.dandielo.citizens.traders_v3.utils.items.Attribute;
 
-@Attribute(name="Wand Auto-Fill", key="wand_fill", priority = 5)
+@Attribute(name="Wand Auto-Fill", key=".wand_fill", priority = 5)
 public class WandFillFlag extends NBTItemFlag {
 	
 	public WandFillFlag(String key) {