+++ src/main/java/com/elmakers/mine/bukkit/traders/WandKeepFlag.java
@@ -2,7 +2,7 @@ package com.elmakers.mine.bukkit.traders;
 
 import net.dandielo.citizens.traders_v3.utils.items.Attribute;
 
-@Attribute(name="Wand Keep on Death", key="wand_keep", priority = 5)
+@Attribute(name="Wand Keep on Death", key=".wand_keep", priority = 5)
 public class WandKeepFlag extends NBTItemFlag {
 	
 	public WandKeepFlag(String key) {