+++ src/main/java/com/elmakers/mine/bukkit/traders/WandOrganizeFlag.java
@@ -2,7 +2,7 @@ package com.elmakers.mine.bukkit.traders;
 
 import net.dandielo.citizens.traders_v3.utils.items.Attribute;
 
-@Attribute(name="Wand Auto-Organize", key="wand_organize", priority = 5)
+@Attribute(name="Wand Auto-Organize", key=".wand_organize", priority = 5)
 public class WandOrganizeFlag extends NBTItemFlag {
 	
 	public WandOrganizeFlag(String key) {