+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
  - Add "keep" and "bound" wand options, for wands that you keep on death and/or are bound to a specific player.
  - Fix dtlTraders integration. Wands, spells and material brushes can be sold by traders. (Requires v 3.1.0, currently a dev build! Thanks again @dandielo !)
- - Add /magic give command, mainly for use with dtlTraders, gives a spell or material brush item.
+ - Add /magic give command, mainly for use with dtlTraders, gives a spell, material brush, or wand upgrade item.
  - Add "allow_cross_world" parameter to Recall spell, the default is to disallow Recall'ing to another world now.
  - Allow selling spells and material brushes on Essentials signs.